Handle messages no longer existing in `remove_subscribed_message`

+++ src/bot.py
@@ -269,9 +269,10 @@ To gain access, you must use `/join` with the password for your group.
             return
         message = await msg_channel.fetch_message(msg.message_id)
 
-        chan_name = message.channel.name if hasattr(message.channel, 'name') else 'unknown channel'
-        print(f'Removing message in {chan_name} from {message.author.name}')
-        await message.delete()  # remove message from discord
+        if message:  # message may have already been deleted manually
+            chan_name = message.channel.name if hasattr(message.channel, 'name') else 'unknown channel'
+            print(f'Removing message in {chan_name} from {message.author.name}')
+            await message.delete()  # remove message from discord
 
         # remove message from subscription list and save to file
         self.subscribed_messages.remove(msg)
